+++ CHANGELOG.md
@@ -12,12 +12,15 @@
 - Added method to get all the selected facets - instance.getAllSelectedFacets()
 - Added a method to get the total count of selected facets - instance.getSelectedFacetsCount()
 - Added a method to get the total count of selected facet values - instance.getSelectedFacetValueCount()
-- Added wrapperTag options to products config
-
+- Added tagName, events and htmlAttributes to each config so that wrapper element created by sdk can be easily customised by customer
+- Added selectedFacets config to manage selected facets, its seperated from facet config
+- Added category-filter support for categoryPath facets, will be removed once backend fix it.
+- Added extra parameters to selected facet template and item template methods
 
 ### Changed
 - fixed the special character issue in multilevel facets.
 - fixed the issue while getting selected multilevel facet method
+- deprecated the selected facets config from facet config.
 
 ## [2.0.0] - 2020-09-28